🍗 Move NavBar + ProfileDropDown components to pages directory
